// src/app/services/models/user.ts
import { Nsi } from "./models";

export class User {
    id: string;
    fullName: string;
    email: string;
    userName: string;
    phoneNumber: string;
    position: string;
    state: string;
    roles: Nsi[];
    roleName: string;
    password: string;
    lastName: string;
    firstName: string;
    patronumic: string;
}